+++ src/hydraulics.h
@@ -11,6 +11,8 @@ double Psi2K(double psi, double Psi_extract, double ws= 3.0);
 NumericVector Psi2K(double psi, NumericVector Psi_extract, double ws= 3.0);
 double averagePsi(NumericVector psi, NumericVector v, double c, double d);
 
+double correctConductanceForViscosity(double kxylem, double temp);
+
 double vanGenuchtenConductance(double psi, double krhizomax, double n, double alpha);
 double xylemConductance(double psi, double kxylemmax, double c, double d);
 double xylemPsi(double kxylem, double kxylemmax, double c, double d);